requirements.txt
Click==7.0
cloudpickle==0.8.1
cycler==0.10.0
dask==1.1.4
decorator==4.4.0
dlib==19.17.0
face-recognition==1.2.3
face-recognition-models==0.3.0
kiwisolver==1.0.1
matplotlib==3.0.3
networkx==2.2
numpy==1.16.2
pandas==0.24.2
Pillow==5.4.1
pkg-resources==0.0.0
pygame==1.9.4
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.2
scikit-image==0.14.2
scipy==1.2.1
six==1.12.0
toolz==0.9.0
torch==1.0.1.post2
torchvision==0.2.2.post3
